requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.1
cycler==0.10.0
Cython==0.29.22
docker-pycreds==0.4.0
gitdb==4.0.5
GitPython==3.1.13
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.35.0
matplotlib==3.3.4
numba==0.52.0
numpy==1.20.1
pandas==1.2.2
pathtools==0.1.2
patsy==0.5.1
Pillow==8.1.0
plotly==4.14.3
promise==2.3
protobuf==3.15.1
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.4.1
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
sktime==0.5.3
smmap==3.0.5
statsmodels==0.12.2
subprocess32==3.5.4
threadpoolctl==2.1.0
#torch @ file:///D:/dev/workspace/tapnet/torch-1.7.1%2Bcu101-cp37-cp37m-win_amd64.whl
#torchaudio @ file:///D:/dev/workspace/tapnet/torchaudio-0.7.2-cp37-none-win_amd64.whl
#torchvision @ file:///D:/dev/workspace/tapnet/torchvision-0.8.2%2Bcpu-cp37-cp37m-win_amd64.whl
tsaug==0.2.1
tslearn==0.5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.10.20
wincertstore==0.2
tensorflow-gpu==2.2.0
keras==2.3.0
scipy
numpy
pandas
scikit-learn>=0.18.2
h5py
matplotlib
